Progress on tree construction & traversal
