+++ Database/ItemInfoDB.cs
@@ -1,8 +1,8 @@
 ﻿using GilGoblin.WebAPI;
+using Microsoft.EntityFrameworkCore;
 using Serilog;
 using System;
 using System.Collections.Generic;
-using System.Data.Entity;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
