+++ packages.config
@@ -1,7 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
   <package id="BouncyCastle" version="1.8.9" targetFramework="net472" />
-  <package id="EntityFramework" version="6.4.4" targetFramework="net472" />
   <package id="Flurl" version="3.0.2" targetFramework="net472" />
   <package id="Flurl.Http" version="3.2.0" targetFramework="net472" />
   <package id="Google.Protobuf" version="3.19.1" targetFramework="net472" />
@@ -10,28 +9,31 @@
   <package id="K4os.Hash.xxHash" version="1.0.7" targetFramework="net472" />
   <package id="Microsoft.Bcl.AsyncInterfaces" version="1.1.1" targetFramework="net472" />
   <package id="Microsoft.Bcl.HashCode" version="1.1.1" targetFramework="net472" />
+  <package id="Microsoft.CSharp" version="4.7.0" targetFramework="net472" />
   <package id="Microsoft.Data.Sqlite" version="3.1.21" targetFramework="net472" />
   <package id="Microsoft.Data.Sqlite.Core" version="3.1.21" targetFramework="net472" />
   <package id="Microsoft.DotNet.PlatformAbstractions" version="3.1.6" targetFramework="net472" />
-  <package id="Microsoft.EntityFrameworkCore" version="3.1.21" targetFramework="net472" />
-  <package id="Microsoft.EntityFrameworkCore.Abstractions" version="3.1.21" targetFramework="net472" />
-  <package id="Microsoft.EntityFrameworkCore.Analyzers" version="3.1.21" targetFramework="net472" />
-  <package id="Microsoft.EntityFrameworkCore.Relational" version="3.1.21" targetFramework="net472" />
+  <package id="Microsoft.EntityFrameworkCore" version="3.1.22" targetFramework="net472" />
+  <package id="Microsoft.EntityFrameworkCore.Abstractions" version="3.1.22" targetFramework="net472" />
+  <package id="Microsoft.EntityFrameworkCore.Analyzers" version="3.1.22" targetFramework="net472" />
+  <package id="Microsoft.EntityFrameworkCore.Design" version="3.1.22" targetFramework="net472" developmentDependency="true" />
+  <package id="Microsoft.EntityFrameworkCore.Relational" version="3.1.22" targetFramework="net472" />
   <package id="Microsoft.EntityFrameworkCore.Sqlite" version="3.1.21" targetFramework="net472" />
   <package id="Microsoft.EntityFrameworkCore.Sqlite.Core" version="3.1.21" targetFramework="net472" />
-  <package id="Microsoft.Extensions.Caching.Abstractions" version="3.1.21" targetFramework="net472" />
-  <package id="Microsoft.Extensions.Caching.Memory" version="3.1.21" targetFramework="net472" />
-  <package id="Microsoft.Extensions.Configuration" version="3.1.21" targetFramework="net472" />
-  <package id="Microsoft.Extensions.Configuration.Abstractions" version="3.1.21" targetFramework="net472" />
-  <package id="Microsoft.Extensions.Configuration.Binder" version="3.1.21" targetFramework="net472" />
-  <package id="Microsoft.Extensions.DependencyInjection" version="3.1.21" targetFramework="net472" />
-  <package id="Microsoft.Extensions.DependencyInjection.Abstractions" version="3.1.21" targetFramework="net472" />
+  <package id="Microsoft.EntityFrameworkCore.Tools" version="3.1.22" targetFramework="net472" developmentDependency="true" />
+  <package id="Microsoft.Extensions.Caching.Abstractions" version="3.1.22" targetFramework="net472" />
+  <package id="Microsoft.Extensions.Caching.Memory" version="3.1.22" targetFramework="net472" />
+  <package id="Microsoft.Extensions.Configuration" version="3.1.22" targetFramework="net472" />
+  <package id="Microsoft.Extensions.Configuration.Abstractions" version="3.1.22" targetFramework="net472" />
+  <package id="Microsoft.Extensions.Configuration.Binder" version="3.1.22" targetFramework="net472" />
+  <package id="Microsoft.Extensions.DependencyInjection" version="3.1.22" targetFramework="net472" />
+  <package id="Microsoft.Extensions.DependencyInjection.Abstractions" version="3.1.22" targetFramework="net472" />
   <package id="Microsoft.Extensions.DependencyModel" version="3.1.6" targetFramework="net472" />
-  <package id="Microsoft.Extensions.Logging" version="3.1.21" targetFramework="net472" />
-  <package id="Microsoft.Extensions.Logging.Abstractions" version="3.1.21" targetFramework="net472" />
-  <package id="Microsoft.Extensions.Options" version="3.1.21" targetFramework="net472" />
+  <package id="Microsoft.Extensions.Logging" version="3.1.22" targetFramework="net472" />
+  <package id="Microsoft.Extensions.Logging.Abstractions" version="3.1.22" targetFramework="net472" />
+  <package id="Microsoft.Extensions.Options" version="3.1.22" targetFramework="net472" />
   <package id="Microsoft.Extensions.Options.ConfigurationExtensions" version="2.0.0" targetFramework="net472" />
-  <package id="Microsoft.Extensions.Primitives" version="3.1.21" targetFramework="net472" />
+  <package id="Microsoft.Extensions.Primitives" version="3.1.22" targetFramework="net472" />
   <package id="MySql.Data" version="8.0.27" targetFramework="net472" />
   <package id="Newtonsoft.Json" version="13.0.1" targetFramework="net472" />
   <package id="Serilog" version="2.10.0" targetFramework="net472" />
